+++ ros/src/computing/planning/common/lib/astar_search/package.xml
