+++ ros/src/computing/planning/motion/packages/waypoint_planner/include/waypoint_planner/astar_avoid/astar_avoid.h
@@ -44,7 +44,7 @@
 #include <autoware_msgs/LaneArray.h>
 
 #include "waypoint_follower/libwaypoint_follower.h"
-#include "astar_search.h"
+#include "astar_search/astar_search.h"
 
 class AstarAvoid
 {
