+++ ros/src/computing/planning/motion/packages/waypoint_planner/src/astar_avoid/astar_avoid_node.cpp
@@ -28,7 +28,7 @@
  *  OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
 
-#include "astar_avoid.h"
+#include "waypoint_planner/astar_avoid/astar_avoid.h"
 
 int main(int argc, char** argv)
 {
