+++ ros/src/computing/planning/motion/packages/waypoint_planner/src/velocity_set/libvelocity_set.cpp
@@ -1,4 +1,4 @@
-#include "libvelocity_set.h"
+#include "../../src/velocity_set/libvelocity_set.h"
 
 // extract edge points from zebra zone
 std::vector<geometry_msgs::Point> removeNeedlessPoints(std::vector<geometry_msgs::Point> &area_points)
