+++ ros/src/computing/planning/motion/packages/waypoint_planner/src/velocity_set/velocity_set_info.cpp
@@ -28,7 +28,7 @@
  *  OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include "velocity_set_info.h"
+#include "../../src/velocity_set/velocity_set_info.h"
 
 void joinPoints(const pcl::PointCloud<pcl::PointXYZ>& points1, pcl::PointCloud<pcl::PointXYZ>* points2)
 {
